refactor: move gesture-custom to utils folder
